pyproject.toml
[tool.poetry]
name = "sosecrets"
version = "0.1.0"
description = "Simple wrapper to secure your secrets."
authors = ["Jim Chng <jimchng@outlook.com>"]
license = "MIT"
readme = "README.md"
keywords = ["secrets", "security"]

[tool.poetry.dependencies]
python = "^3.8"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"


[tool.poetry.group.examples.dependencies]
cryptography = "^40.0.2"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.bumpver]
current_version = "1.0.0"
version_pattern = "MAJOR.MINOR.PATCH"
commit_message = "bump version {old_version} -> {new_version}"
commit = true
tag = true
push = true

[tool.bumpver.file_patterns]
"pyproject.toml" = [
    'current_version = "{version}"',
    'version = "{version}"'
]

"src/sosecrets/__init__.py" = [
    'version = "{version}"'
]

"README.md" = [
    "Version: {version}",
]

"docs/README.md" = [
    "Version: {version}",
]
